backend/requirements.txt
adb_shell==0.4.4
adblockparser==0.7
aider-chat==0.74.2
aider-install==0.1.3
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alembic==1.13.2
altair==4.2.2
annotated-types==0.7.0
anthropic==0.47.2
anyio==4.8.0
Appium-Python-Client==4.3.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.1.0
autocommand==2.2.2
autopep8==2.0.4
backoff==2.2.1
backports.tarfile==1.2.0
bcrypt==4.1.3
beautifulsoup4==4.13.3
black==23.3.0
bleach==6.1.0
blinker==1.7.0
bottle==0.13.2
bs4==0.0.2
cachetools==5.3.3
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
cheroot==10.0.1
CherryPy==18.10.0
cherrypy-cors==1.7.0
chroma-hnswlib==0.7.2
chromadb==0.4.6
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.2.1
cryptography==3.4.8
cycler==0.12.1
dataclasses-json==0.5.7
dataclasses-json-speakeasy==0.5.11
datasets==2.21.0
debugpy==1.8.12
decorator==5.2.1
defusedxml==0.7.1
demogpt @ file:///Users/harrytowers/Documents/Development/AI/DemoGPT
deprecation==2.1.0
diff-match-patch==20241021
dill==0.3.8
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
emoji==2.10.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==2.2.0
ExifRead==2.3.2
fastapi==0.99.1
fastjsonschema==2.19.1
filelock==3.17.0
filetype==1.2.0
flake8==7.1.1
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.7
fonttools==4.53.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.2.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
gpt-engineer==0.3.1
greenlet==3.0.3
grep-ast==0.5.0
h11==0.14.0
httpagentparser==1.9.5
httpcore==1.0.7
httptools==0.6.1
httpx==0.28.1
huggingface-hub==0.28.1
humanfriendly==10.0
identify==2.5.35
idna==3.10
ignorant==1.2
importlib_metadata==7.2.1
importlib_resources==6.5.2
inflect==7.3.1
ipaddr==2.2.0
ipwhois==1.1.0
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.collections==5.0.1
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.14.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.3.2
json5==0.10.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keyring==25.6.0
kiwisolver==1.4.5
langchain==0.2.16
langchain-anthropic==0.1.23
langchain-community==0.2.17
langchain-core==0.2.40
langchain-experimental==0.0.54
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langdetect==1.0.9
langsmith==0.1.120
lightgbm==4.6.0
litellm==1.60.6
llama_cpp_python==0.2.56
lxml==4.9.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.20.2
marshmallow-enum==1.5.1
matplotlib==3.9.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
mixpanel==4.10.1
monotonic==1.6
more-itertools==10.3.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
multitasking==0.0.11
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.2
nbformat==5.10.3
nest-asyncio==1.6.0
netaddr==0.10.1
networkx==3.2.1
nltk==3.8.1
nodeenv==1.8.0
numerapi==2.20.0
numexpr==2.9.0
numpy==1.26.4
onnxruntime==1.17.1
openai==1.65.1
opencv-python==4.11.0.86
openpyxl==3.1.5
orjson==3.9.15
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
paramiko==3.4.0
parso==0.8.4
pathspec==0.12.1
patsy==1.0.1
pdf2image==1.17.0
pdfminer.six==20231228
peewee==3.17.9
pexpect==4.9.0
phonenumbers==8.13.42
pillow==10.4.0
platformdirs==4.2.0
playwright==1.41.2
ply==3.11
portend==3.2.0
posthog==3.11.0
pre-commit==3.6.2
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==4.25.3
proxy_tools==0.1.0
proxyscrape==0.3.0
psutil==5.9.0
ptyprocess==0.7.0
publicsuffixlist==0.9.4
pulsar-client==3.4.0
pure_eval==0.2.3
pyarrow==15.0.1
pyasn1==0.6.0
PyAutoGUI==0.9.54
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==1.10.21
pydantic_core==2.27.2
pydeck==0.8.1b0
pydub==0.25.1
pyee==11.0.1
pyflakes==3.2.0
PyGetWindow==0.0.9
pygexf==0.2.2
Pygments==2.19.1
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyngrok==7.1.6
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-Quartz==11.0
pyobjc-framework-Security==11.0
pyobjc-framework-WebKit==11.0
pyOpenSSL==21.0.0
pypandoc==1.15
pyparsing==3.1.2
PyPDF2==1.28.6
pyperclip==1.9.0
PyPika==0.48.9
PyRect==0.2.0
PyScreeze==1.0.1
pysmi==1.4.4
pysnmp==4.4.6
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.1
python-iso639==2024.2.7
python-magic==0.4.27
python-multipart==0.0.20
python-pptx==0.6.23
python-whois==0.7.3
pytweening==1.2.0
pytz==2024.1
pywebview==5.4
PyYAML==6.0.2
pyzmq==25.1.2
rapidfuzz==3.6.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.2
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
rubicon-objc==0.5.0
rudder-sdk-python==2.1.3
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
secure==0.3.0
selenium==4.26.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.6
SpeechRecognition==3.10.4
SQLAlchemy==2.0.28
stack-data==0.6.3
starlette==0.27.0
statsmodels==0.14.4
streamlit==1.32.1
sympy==1.12
tabulate==0.9.0
tavily-python==0.3.3
tempora==5.6.0
tenacity==8.2.3
termcolor==2.3.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.2.1
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.5
toolz==0.12.1
tornado==6.4
tqdm==4.67.1
traitlets==5.14.2
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
trio==0.25.0
trio-websocket==0.11.1
typeguard==4.3.0
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
unstructured==0.12.6
unstructured-client==0.18.0
urllib3==2.3.0
uv==0.6.1
uvicorn==0.28.0
uvloop==0.19.0
virtualenv==20.25.1
watchfiles==1.0.4
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
XlsxWriter==3.2.0
xxhash==3.5.0
yarl==1.18.3
yfinance==0.2.54
youtube-transcript-api==0.6.2
zc.lockfile==3.0.post1
zipp==3.21.0
